Tracked down the differences between the stress in the ancient CP and in the recent CVS. For a small test, the stress for USPPs now agrees much better with the ancient CP code. All the interested parties please test.
 For the fix, backported the old CP's "bess" routine, since the logic
for "sph_bes" was the cause of the differences. Please be careful when
making "sph_bes" work for tpre=.true.

 Another puzzle - converged CP energies were identical from the
ancient CP through CP in PWSCF v2.1, but now are slightly different
(same machine, compiler, etc.) The biggest difference is the
exchange-correlation energy. Does anyone know what is the EXACT reason
for this?

 Kostya


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@3046 c92efa57-630b-4861-b058-cf58834340f0