Normalization of atomic wavefuncitons is checked (only for those with non-negative occupation) after pseudopotential reading and if it found to be different from 1 by more than 1.d-6 the wavefunction is renormalized.
Should make no difference in all usual cases since atomic wfcs are only
used in order to generate the initial set of trial wfcs and normalization
is not important there...

It is VERY important for LDA+U calculations using atomic wfc in the progector
(the default case) since proper normalization is assumed and lack of it leads
to disasters.

Hopefully this will solve many of the difficulties people have encounterd
when using LDA+U with older RRKJ pseudopotentials, where normalization
was not imposed properly (and nobody realizes except when doing LDA+U) or
pseudopotentials of unknown origin.


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@5054 c92efa57-630b-4861-b058-cf58834340f0