Small touches for easier compilation in MACOS X 10.6 This one enables compilation of pw using macports g95 gcc and apple libtool I still have to test if the compilation was proper.
MacOs X 10.6 is an esoteric operating system, especially how the '64 bit' is
handled. Any help on the matter is really appreciated

1) g95 from macports complain that beta in exx.f90 is not an integer. I did not
 see any reason why it was not an integer either. Please verify

2) apple suplied ranlib has some 'touches' related to 'universal' libraries
which mix 64 bit 32 bit and ppc all together, this results in lost symbols
in some cases (sadly Q-E is one such case). Details at
http://osdir.com/ml/gnu.libtool.general/2002-07/msg00045.html
modifying configure.ac such that in mac pcs ranlib is run in compatibility mode

The next patch will be about MKL. At least in the trial version, MKL is supplied
with ifort by default, however configure does not look for it. I will check
if it is due to some esoteric behaviour, otherwise enable it.


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@6692 c92efa57-630b-4861-b058-cf58834340f0