The internal blas was not compiling correctly due to different compile flags used.
This is my first attempt to fix.

The solution I use is to define a new "precious" variable (as defined by gnu
automake standarts) extlib_flags, and pass it to make_blas.inc.in
The advantage of defining a precious variable is mostly due to caching against
problems in automatic remaking. The document is here:
http://www.gnu.org/software/hello/manual/autoconf/Setting-Output-Variables.html#Setting-Output-Variables

I also suggest setting variables from ln:145 to ln:151 in configure.ac as
precious variables to be on the safe side and conform to GNU guidelines.


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@7154 c92efa57-630b-4861-b058-cf58834340f0