With this commit I include the possibility of running a phonon calculation with a different k-point grid than that used for the charge density. This is equivalent to what was present in previous QE versions when pw.x was run with the option calculation="phonon" and a number of k-points different from that used for the charge density.
To use this option specify 
nk1, nk2, nk3 with values differents from zero in the input file
of the ph.x
The variables k1, k2, k3 can also be used to shift the regular
grid from the origin.
An example of the ph.x input file is

Ph
 &inputph
    prefix='MgB2',
    amass(1)= 10.811,    
    amass(2)= 24.305,    
    fildyn='dynq1',
    fildvscf='dvscf_1',
    outdir='./tmpdisp_nk/',
    tr2_ph=1.0d-14,
    trans=.true.,
    epsil=.false.,
    ldisp=.true.,
    nq1=2,nq2=2,nq3=2,
    nk1=16,               <-------------------
    nk2=16,               <-------------------
    nk3=12,               <-------------------
    k1=0,
    k2=0,
    k3=0,
 &end 

The defaults values for k1,k2,k3 and nk1, nk2, nk3 is zero.
The option works both with ldisp=.true. or with the calculation
for a fixed q.
See doc-def/INPUT_PH.def for details on the input file.



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@7744 c92efa57-630b-4861-b058-cf58834340f0