Module mp_global split into more modules, one per parallelization level. Module mp_global is still there for compatibility and still contains the routine mp_startup that performs the intialization, but shouldn't be any longer used in new developments and should be slowly replaced by specific modules. New module containing command-line options added. Command-line options are read at the beginning of the run and broadcast to all processors (there is no guarantee that command-line options cam be accessed by all mpi  processes). Nothing should be broken by these changes but please verify if all parallelization levels work (I have limited access to parallel machines right now)
The reason for these changes, and for those that will come soon, is to make
image parallelization easier (in particular in NEB but also for other cases). 
Right now it is a pain, in part because the initialization is inadequate,
in part because it is never clear who reads/writes what from/to where.
It will take a few weeks before everything converges to a stable state.
Meanwhile, please be patient and fix/report problems.



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@9876 c92efa57-630b-4861-b058-cf58834340f0