A significant optimization of the RAM use has been done for turbo_davison.
For NCPP, memory requirement is reduced to 1/3 ;
while for USPP, it's reduced to 2/3.

There's another option to reduce the ram use of USPP by another factor of 2 by
setting poor_of_ram=.true., so it requires the same ram as NCPP. The price that
one has to pay is to increase the time of calculation because one has to calculate
s*basis every time one needs it instead of storing all the s*basis vectors. The
induced calcualtion can be neglectable when the size of subspace is small,
but it increases rapidly when the size of subspace is big. For example, when the
size of subspace is about 200, the cal_bec and s_psi calculation will spend more 
than half of the whole calculation time. So think it three times before deciding 
how to choose.



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@10316 c92efa57-630b-4861-b058-cf58834340f0