ld1inc.f90 : a small bug was preventing the generation of non-PAW pseudopotentials with GIPAw reconstruction data
ld1.f90    : running the pseudopotential test when also including GIPAW reconstruction data is not advisable, as the test configuration may not make sense and produce random and unrelated error message. The possible workaround is to test the pseudopotential first and only include the gipaw data later. A real solution would be to not recycle the test configuration to do other stuff.

LP



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@11296 c92efa57-630b-4861-b058-cf58834340f0