band group parallelization slightly modified to make it more flexible, and little more efficient.
 subroutine init_index_over_band ( comm, nbnd ) that set ibnd_start and ibnd_end
variables requiring comm=inter_bgrp_comm is removed and replaced by
 subroutine set_bgrp_indices ( nbnd, ibnd_start, ibnd_end ) implementing the same
relationships between its arguments but:
 
 - forcing the use of inter_bgrp_comm from the same mp_bands module,

 - returning ibnd_start and ibnd_end as explicit outputs that are not anymore kept
in the module. In this way other quantities can be distributes if needed in any 
given routine without too many non-local effects.

 For compatibility with TDDFPT, that uses the bgrp parallelization and loads
ibnd_start/ibnd_end trhough mp_global module, these two variables are moved in 
a dedicated module mp_bands_TDDFPT included in Module/mp_bands.f90. This is done
to avoid too much invasive changes in a code i don't know well. In this way the 
needed changes are very localized and transparent, the code compiles correctly 
so I think it should work exactly as before. 
In my opinion the two variables should be moved somewhere inside TDDFPT.

 Band parallelization is extended to h_psi(lda,n,m,psi,hpsi) and s_psi routines 
(only when .not.exx_is_active because otherwise it is already used inside vexx)
for generic values of m (of course it gives a speedup only when m is not too small
compared to nbgrp but it works also if m < nbgrp ).

 Compatibility with task groups has not be explored but should not be conceptually
different from how it works in the exx case.


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@11835 c92efa57-630b-4861-b058-cf58834340f0