band parallelization works also with real_sapce at gamma_only
Tested on example02/co.rx.in (cutoffs increased to 60/360 to improve accuracy of real space 
integrals, still not converged tho): the serial and parallel verisions give the same results, 
however the forces on atoms do not seem to be very accurate as the code is unable to relax 
the simple dimer. I don't know if this is due to lack of ecutwfc/ecutrho convergence or
indicates some deeper problem in the real_space treatment. 

In order to make bgrp parallelization work sum_bec routine in sum_band has been corrected 
so that invfft_orbital_gamma and calbec_rs_gamma are called as

       call invfft_orbital_gamma(evc,ibnd,ibnd_end) ! was ...,nbnd)
       call calbec_rs_gamma(ibnd,ibnd_end,becp%r)   ! was ...,nbnd,becp%r)

using ibnd_end and not this_bgrp_nbnd because, contrary to what was stated in the routines headers,
the corresponding input parameter is the index of the last band to be considered and not the total
number of bands to be calculated. They have often, but not always, the same value.
Comments and variable names in realus.f90 updated to reflect this fact. 

minor cleanup of unused variables in h_1psi.f90



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@11868 c92efa57-630b-4861-b058-cf58834340f0