print cell and tau to 9 dp in pw output
Input structure is only printed as a combination of celldm,
alat crystal axes and tau. The combination of these mean that
determing the structure from the output alone is limited to about
5 or 6 decimal places of precision compared to 9 dp for the
ATOMIC_POSITIONS card format. It is possible to print 9 or 10 dp
with some slight shuffling of fields stil keeping within 80 character
output.

This is useful for better precision in a pw output parser that
I wrote for ase.